refactor: split api.py by domain
