+++ snap_http/__init__.py
@@ -19,9 +19,6 @@ from .api import (
     snapshots,
     forget_snapshot,
     save_snapshot,
-    snapshots,
-    forget_snapshot,
-    save_snapshot,
     switch,
     switch_all,
     unhold,
